+++ packages/larva/__tests__/_test-boilerplate.js
@@ -2,7 +2,7 @@ const assert = require( 'assert' );
 const path = require( 'path' );
 const fixture = path.join( __dirname, '../../fixtures' );
 
-// If needed:
+// If needed, will return projectPatternsDir pointing to fixtures:
 // const getAppConfiguration = require( '../../../lib/utils/getAppConfiguration' );
 
 describe( 'describe the test', () => {