+++ packages/larva/lib/utils/getSubDirectoryNames.js
@@ -3,11 +3,12 @@ const chalk = require( 'chalk' );
 
 /**
  * Get Subdirectory names
- * 
- * @para path String of path to subdirectory
- * 
+ *
+ * @param {string} path path to a directory containing subdirectories.
+ * @return {array} array of strings representing names of subdirectories.
+ *
  * @link: https://stackoverflow.com/questions/18112204/get-all-directories-within-directory-nodejs
- **/ 
+ **/
 
 module.exports = function getSubDirectoryNames( path ) {
 	try {
@@ -15,6 +16,6 @@ module.exports = function getSubDirectoryNames( path ) {
 			return fs.statSync( path + '/' + file ).isDirectory();
 		});
 	} catch {
-		console.warn( chalk.yellow( path + ' doesn\'t exist.' ) );
+		console.error( path + ' doesn\'t exist.' );
 	}
 }